requirements.txt
Flask==2.1.0
face_recognition==1.3.0
opencv-python-headless==4.5.3.56
Pillow==8.4.0
requests==2.26.0
numpy==1.21.2
